refactor:game.html => index.html
